requirements.txt
Shapely==1.7.1
numpy==1.20.3
scipy==1.7.3
pydicom==2.1.2
matplotlib==3.5.0
scikit_image==0.18.3
Pillow==9.3.0
skimage==0.0
